# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.FileResponse:
    properties:
      content_base64:
        example: MQ0KMg0KMw0KdGVzdCBzdWNjZXNzISE=
        type: string
      filename:
        example: test.txt
        type: string
    type: object
  models.GetFileAnonymousRequest:
    properties:
      file_path:
        example: /data/test.txt
        type: string
      host:
        example: 192.168.1.1
        type: string
      port:
        example: "22"
        type: string
    type: object
  models.GetFileWithAuthRequest:
    properties:
      file_path:
        example: /data/test.txt
        type: string
      host:
        example: 192.168.1.1
        type: string
      password:
        example: password
        type: string
      port:
        example: "22"
        type: string
      username:
        example: admin
        type: string
    type: object
  models.SendFileRequest:
    properties:
      content_base64:
        example: SGVsbG8sIEZUUCE=
        type: string
      filename:
        example: example.txt
        type: string
      host:
        example: 192.168.30.92
        type: string
      password:
        example: password
        type: string
      port:
        example: "21"
        type: string
      target_path:
        example: upload/
        type: string
      username:
        example: tester
        type: string
    required:
    - content_base64
    - filename
    - host
    - port
    - target_path
    type: object
  swagger.IncorrectFormatError:
    properties:
      response:
        properties:
          code:
            example: 400
            type: integer
          message:
            example: Неверный формат запроса
            type: string
        type: object
      status:
        example: IncorrectFormatError
        type: string
    type: object
  swagger.InternalServerError:
    properties:
      response:
        properties:
          code:
            example: 500
            type: integer
          message:
            example: Внутренняя ошибка сервера
            type: string
        type: object
      status:
        example: InternalServerError
        type: string
    type: object
  swagger.NotFoundError:
    properties:
      response:
        properties:
          code:
            example: 404
            type: integer
          message:
            example: Данные не найдены
            type: string
        type: object
      status:
        example: NotFoundError
        type: string
    type: object
  swagger.SendResponse:
    properties:
      message:
        example: MSG
        type: string
      status:
        example: ok
        type: string
      type:
        example: Empty
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API для работы с протоколом FTP
  title: FTP Client API
  version: 1.0.0
paths:
  /get/anon:
    post:
      consumes:
      - application/json
      description: Подключается к FTP-серверу как anonymous и скачивает указанный
        файл
      parameters:
      - description: Адрес сервера и путь к файлу
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.GetFileAnonymousRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Файл успешно загружен
          schema:
            $ref: '#/definitions/models.FileResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/swagger.IncorrectFormatError'
        "404":
          description: Файл не найден
          schema:
            $ref: '#/definitions/swagger.NotFoundError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/swagger.InternalServerError'
      summary: Скачать файл с FTP (анонимный доступ)
      tags:
      - FTP_Anon
  /get/auth:
    post:
      consumes:
      - application/json
      description: Подключается к FTP-серверу с использованием логина и пароля, и
        скачивает указанный файл
      parameters:
      - description: Параметры подключения и файл
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.GetFileWithAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Файл успешно загружен
          schema:
            $ref: '#/definitions/models.FileResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/swagger.IncorrectFormatError'
        "404":
          description: Файл не найден
          schema:
            $ref: '#/definitions/swagger.NotFoundError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/swagger.InternalServerError'
      summary: Скачать файл с FTP (авторизация)
      tags:
      - FTP_Auth
  /send/anon:
    post:
      consumes:
      - application/json
      description: Подключается к FTP-серверу как anonymous и загружает файл
      parameters:
      - description: Данные подключения и файл
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SendFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Файл успешно отправлен
          schema:
            $ref: '#/definitions/swagger.SendResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/swagger.IncorrectFormatError'
        "500":
          description: Ошибка при загрузке файла
          schema:
            $ref: '#/definitions/swagger.InternalServerError'
      summary: Загрузить файл на FTP (анонимный доступ)
      tags:
      - FTP_Anon
  /send/auth:
    post:
      consumes:
      - application/json
      description: Подключается к FTP-серверу с логином и паролем и загружает указанный
        файл
      parameters:
      - description: Данные подключения и файл
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SendFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Файл успешно отправлен
          schema:
            $ref: '#/definitions/swagger.SendResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/swagger.IncorrectFormatError'
        "500":
          description: Ошибка при загрузке файла
          schema:
            $ref: '#/definitions/swagger.InternalServerError'
      summary: Загрузить файл на FTP (авторизация)
      tags:
      - FTP_Auth
swagger: "2.0"